fix multi sample
