+++ pkg/sql/colexec/sample/samplePool2.go
@@ -15,8 +15,11 @@
 package sample
 
 import (
+	"github.com/matrixorigin/matrixone/pkg/common/moerr"
 	"github.com/matrixorigin/matrixone/pkg/common/mpool"
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 )
 
@@ -29,6 +32,7 @@ type poolData struct {
 	invalidBatch *batch.Batch
 }
 
+// maybe helpful for code performance sometimes.
 func (pd *poolData) needAppendInvalid() bool {
 	return pd.invalidBatch == nil
 }
@@ -72,7 +76,70 @@ func (pd *poolData) replaceValidRow(mp *mpool.MPool, bat *batch.Batch, row1, row
 	return batRowReplace(mp, pd.validBatch, bat, row1, row2)
 }
 
+// batRowReplace replaces the row1 of toBatch with the bat's row2.
+// TODO: need an optimized function to do the row replace work.
+func batRowReplace(mp *mpool.MPool, toBatch *batch.Batch, bat *batch.Batch, row1, row2 int) (err error) {
+	var right int
+	for i, vec := range bat.Vecs {
+		right = row2
+		if vec.IsConst() {
+			right = 0
+		}
+		switch vec.GetType().Oid {
+		case types.T_int8:
+			err = vector.SetFixedAt[int8](toBatch.Vecs[i], row1, vector.GetFixedAt[int8](vec, right))
+		case types.T_int16:
+			err = vector.SetFixedAt[int16](toBatch.Vecs[i], row1, vector.GetFixedAt[int16](vec, right))
+		case types.T_int32:
+			err = vector.SetFixedAt[int32](toBatch.Vecs[i], row1, vector.GetFixedAt[int32](vec, right))
+		case types.T_int64:
+			err = vector.SetFixedAt[int64](toBatch.Vecs[i], row1, vector.GetFixedAt[int64](vec, right))
+		case types.T_uint8:
+			err = vector.SetFixedAt[uint8](toBatch.Vecs[i], row1, vector.GetFixedAt[uint8](vec, right))
+		case types.T_uint16:
+			err = vector.SetFixedAt[uint16](toBatch.Vecs[i], row1, vector.GetFixedAt[uint16](vec, right))
+		case types.T_uint32:
+			err = vector.SetFixedAt[uint32](toBatch.Vecs[i], row1, vector.GetFixedAt[uint32](vec, right))
+		case types.T_uint64:
+			err = vector.SetFixedAt[uint64](toBatch.Vecs[i], row1, vector.GetFixedAt[uint64](vec, right))
+		case types.T_float32:
+			err = vector.SetFixedAt[float32](toBatch.Vecs[i], row1, vector.GetFixedAt[float32](vec, right))
+		case types.T_float64:
+			err = vector.SetFixedAt[float64](toBatch.Vecs[i], row1, vector.GetFixedAt[float64](vec, right))
+		case types.T_date:
+			err = vector.SetFixedAt[types.Date](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Date](vec, right))
+		case types.T_datetime:
+			err = vector.SetFixedAt[types.Datetime](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Datetime](vec, right))
+		case types.T_timestamp:
+			err = vector.SetFixedAt[types.Timestamp](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Timestamp](vec, right))
+		case types.T_time:
+			err = vector.SetFixedAt[types.Time](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Time](vec, right))
+		case types.T_enum:
+			err = vector.SetFixedAt[types.Enum](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Enum](vec, right))
+		case types.T_decimal64:
+			err = vector.SetFixedAt[types.Decimal64](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Decimal64](vec, right))
+		case types.T_decimal128:
+			err = vector.SetFixedAt[types.Decimal128](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Decimal128](vec, right))
+		case types.T_TS:
+			err = vector.SetFixedAt[types.TS](toBatch.Vecs[i], row1, vector.GetFixedAt[types.TS](vec, right))
+		case types.T_Rowid:
+			err = vector.SetFixedAt[types.Rowid](toBatch.Vecs[i], row1, vector.GetFixedAt[types.Rowid](vec, right))
+		case types.T_char, types.T_varchar, types.T_binary, types.T_varbinary,
+			types.T_json, types.T_blob, types.T_text,
+			types.T_array_float32, types.T_array_float64:
+			err = vector.SetBytesAt(toBatch.Vecs[i], row1, vec.GetBytesAt(right), mp)
+		default:
+			err = moerr.NewInternalErrorNoCtx("unsupported type for sample pool.")
+		}
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 // output1 returns the result of poolData and set the source pointer to be nil.
+// priority: validBatch > invalidBatch.
 func (pd *poolData) output() (bat *batch.Batch) {
 	if pd.validBatch != nil {
 		bat = pd.validBatch
